Revert "move `server.ts` from `utils/` to `common/`"
This reverts commit 1beb0ad7a22744535c12a4b8c78951c20e0315f9.
